refactor(core): merge rules/performance/ into rules/perf/
All PERF rules now live in one directory. No behavior change — same rule
ids, same allRules contents, same public exports; only the source files'
location and two import paths in rules/index.ts changed.
