+++ AGENTS.md
@@ -40,7 +40,7 @@ CI (`.github/workflows/ci.yml`) runs four jobs: `lint`, `check` (build + typeche
   - `fix(cli): make --diff/--staged work when the project is not at the git repo root`
   - `test(cli): pin behavior for malformed .svelte files in both passes`
   - Other prefixes in use: `feat(vite):`, `docs:`, `chore:`.
-- **Adding a rule**: create `packages/core/src/rules/<category>/xxxNNN-slug.ts`, then register it in **four** places: `packages/core/src/rules/index.ts` (the import, the `allRules` array, and the re-export block) _and_ `packages/core/src/index.ts`'s own `export { ... } from './rules/index.js'` list, which duplicates the same names. TypeScript won't catch a missed spot in the fourth place (it's a plain re-export list), so grep for the previous rule's id after adding a new one. Add rule docs under `docs/src/content/docs/rules/<id>.md` (en) and `docs/src/content/docs/ja/rules/<id>.md` (ja) — `packages/cli/test/docs-links.test.ts` fails the build if either is missing. Note: for historical reasons performance rules are split across `rules/perf/` (PERF001–008) and `rules/performance/` (PERF009–010) — check both when looking for an existing PERF rule.
+- **Adding a rule**: create `packages/core/src/rules/<category>/xxxNNN-slug.ts`, then register it in **four** places: `packages/core/src/rules/index.ts` (the import, the `allRules` array, and the re-export block) _and_ `packages/core/src/index.ts`'s own `export { ... } from './rules/index.js'` list, which duplicates the same names. TypeScript won't catch a missed spot in the fourth place (it's a plain re-export list), so grep for the previous rule's id after adding a new one. Add rule docs under `docs/src/content/docs/rules/<id>.md` (en) and `docs/src/content/docs/ja/rules/<id>.md` (ja) — `packages/cli/test/docs-links.test.ts` fails the build if either is missing.
 - **Tests**: vitest, per-package `test/` directories; fixtures live under `test/fixtures/`.
 
 ## Design docs
