+++ packages/core/src/rules/perf/perf009-heavy-import.ts
