+++ packages/core/src/rules/perf/perf010-namespace-import.ts
